Cretae DisplayABook.js for a single book
