--- src/components/Categories.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-class Categories extends React.Component {
-  render() {
-    return (
-      <div>
-        <button>Check Status</button>
-
-      </div>
-    );
-  }
-}
-
-export default Categories;
